+++ CHANGELOG.md
@@ -22,6 +22,7 @@ title: Changelog
 * `<cc-env-var-form>`: fix toggling to JSON mode while in skeleton state.
 * `<cc-badge>`: add skeleton mode
 * `<cc-input-text>`: remove the unique id generation technique and rely on Shadow DOM isolation instead.
+* `<cc-input-number>`: remove the unique id generation technique and rely on Shadow DOM isolation instead.
 * New component:
   * `<cc-action-dispatcher>`